v2.6.9-dev
1.修改串口接收机制，修改电机停止bug
2.修改电量超过100的情况
3.增加当电机运行时间为0xFF时永久运行
4.增加新协议版本号为三个字节
5.禁止mcu进入休眠

+++ user/user_timer.c
@@ -5,7 +5,7 @@ uint16_t soft_timer_t[TIMER_MAX] = {0};
 
 
 //设置中断时间
-void set_soft_timer(SOFT_TIMER index, uint16_t nms)
+void set_soft_timer(SOFT_TIMER index, uint32_t nms)
 {		
 		soft_timer[index] = nms;
 		soft_timer_t[index]  = nms;	 //防内在泄漏纠正
